Couple printer enabled to button color
